requirements.txt
numpy
librosa
pandas
noisereduce
scikit-image
matplotlib
tqdm
scipy
torch
tensorflow
Pillow